Fix: 정답률 로그 Int형으로 변환

See merge request s11-bigdata-dist-sub1/S11P21B206!138

+++ backend/log-service/src/main/java/com/helloword/logservice/domain/log/dto/response/GameStatsResponseDto.java
@@ -7,20 +7,20 @@ import lombok.Getter;
 @AllArgsConstructor
 public class GameStatsResponseDto {
 
-	private Double speedGameKidAverageCorrectRate;
-	private Double speedGameGlobalAverageCorrectRate;
+	private Integer speedGameKidAverageCorrectRate;
+	private Integer speedGameGlobalAverageCorrectRate;
 	private Integer speedGameKidAveragePlayTime;
 	private Integer speedGameGlobalAveragePlayTime;
-	private Double speechGameKidAverageCorrectRate;
-	private Double speechGameGlobalAverageCorrectRate;
+	private Integer speechGameKidAverageCorrectRate;
+	private Integer speechGameGlobalAverageCorrectRate;
 	private Integer speechGameKidAveragePlayTime;
 	private Integer speechGameGlobalAveragePlayTime;
-	private Double pairGameKidAverageCorrectRate;
-	private Double pairGameGlobalAverageCorrectRate;
+	private Integer pairGameKidAverageCorrectRate;
+	private Integer pairGameGlobalAverageCorrectRate;
 	private Integer pairGameKidAveragePlayTime;
 	private Integer pairGameGlobalAveragePlayTime;
-	private Double fairytaleGameKidAverageCorrectRate;
-	private Double fairytaleGameGlobalAverageCorrectRate;
+	private Integer fairytaleGameKidAverageCorrectRate;
+	private Integer fairytaleGameGlobalAverageCorrectRate;
 	private Integer fairytaleGameKidAveragePlayTime;
 	private Integer fairytaleGameGlobalAveragePlayTime;
 }
